Move Types Into New Structure
