+++ @nerve/domains/blog/index.d.ts
@@ -1,0 +1,9 @@
+import { FixedImageProps } from '@nerve/core/components';
+import { PrismicExternalLink } from '@nerve/shared/types';
+
+export interface Author {
+    name: string;
+    bio: string;
+    instagram: PrismicExternalLink;
+    headshot: FixedImageProps;
+}
