+++ @nerve/shared/types/index.d.ts
@@ -1,0 +1,2 @@
+export * from './gatsby'
+export * from './prismic'
